Move BAGL UI code for SIGN_MESSAGE instruction in dedicated file
