+++ src/handle_sign_message.c
@@ -1,20 +1,18 @@
-#include "os.h"
 #include "io.h"
-#include "ux.h"
-#include "cx.h"
 #include "menu.h"
 #include "utils.h"
+#include "handle_swap_sign_transaction.h"
+
 #include "sol/parser.h"
 #include "sol/printer.h"
 #include "sol/print_config.h"
 #include "sol/message.h"
 #include "sol/transaction_summary.h"
-#include "globals.h"
-#include "apdu.h"
 
-#include "handle_swap_sign_transaction.h"
+#include "handle_sign_message.h"
 
-static uint8_t set_result_sign_message() {
+
+uint8_t set_result_sign_message(void) {
     uint8_t signature[SIGNATURE_LENGTH];
     cx_ecfp_private_key_t privateKey;
     BEGIN_TRY {
@@ -44,46 +42,6 @@ static uint8_t set_result_sign_message() {
     END_TRY;
     return SIGNATURE_LENGTH;
 }
-
-//////////////////////////////////////////////////////////////////////
-
-UX_STEP_CB(ux_approve_step,
-           pb,
-           sendResponse(set_result_sign_message(), true, true),
-           {
-               &C_icon_validate_14,
-               "Approve",
-           });
-UX_STEP_CB(ux_reject_step,
-           pb,
-           sendResponse(0, false, true),
-           {
-               &C_icon_crossmark,
-               "Reject",
-           });
-UX_STEP_NOCB_INIT(ux_summary_step,
-                  bnnn_paging,
-                  {
-                      size_t step_index = G_ux.flow_stack[stack_slot].index;
-                      enum DisplayFlags flags = DisplayFlagNone;
-                      if (N_storage.settings.pubkey_display == PubkeyDisplayLong) {
-                          flags |= DisplayFlagLongPubkeys;
-                      }
-                      if (transaction_summary_display_item(step_index, flags)) {
-                          THROW(ApduReplySolanaSummaryUpdateFailed);
-                      }
-                  },
-                  {
-                      .title = G_transaction_summary_title,
-                      .text = G_transaction_summary_text,
-                  });
-
-#define MAX_FLOW_STEPS                                     \
-    (MAX_TRANSACTION_SUMMARY_ITEMS + 1 /* approve */       \
-     + 1                               /* reject */        \
-     + 1                               /* FLOW_END_STEP */ \
-    )
-ux_flow_step_t const *flow_steps[MAX_FLOW_STEPS];
 
 static int scan_header_for_signer(const uint32_t *derivation_path,
                                   uint32_t derivation_path_length,
@@ -219,18 +177,7 @@ void handle_sign_message_ui(volatile unsigned int *flags) {
                 THROW(ApduReplySolanaSummaryFinalizeFailed);
             }
         } else {
-            MEMCLEAR(flow_steps);
-            size_t num_flow_steps = 0;
-
-            for (size_t i = 0; i < num_summary_steps; i++) {
-                flow_steps[num_flow_steps++] = &ux_summary_step;
-            }
-
-            flow_steps[num_flow_steps++] = &ux_approve_step;
-            flow_steps[num_flow_steps++] = &ux_reject_step;
-            flow_steps[num_flow_steps++] = FLOW_END_STEP;
-
-            ux_flow_init(0, flow_steps, NULL);
+            start_sign_tx_ui(num_summary_steps);
         }
     } else {
         THROW(ApduReplySolanaSummaryFinalizeFailed);
