+++ src/handle_sign_message.h
@@ -1,12 +1,9 @@
-#include "os.h"
-#include "cx.h"
-#include "globals.h"
+#pragma once
 
-#ifndef _SIGN_MESSAGE_H_
-#define _SIGN_MESSAGE_H_
+uint8_t set_result_sign_message(void);
 
 void handle_sign_message_parse_message(volatile unsigned int *tx);
 
 void handle_sign_message_ui(volatile unsigned int *flags);
 
-#endif
+void start_sign_tx_ui(size_t num_summary_steps);
